+++ leshan-core/src/main/java/org/eclipse/leshan/core/json/jackson/JsonArrayEntrySerDes.java
@@ -95,7 +95,7 @@ public class JsonArrayEntrySerDes extends JacksonJsonSerDes<JsonArrayEntry> {
 
         JsonNode t = o.get("t");
         if (t != null && t.isNumber())
-            jae.setTime(t.floatValue());
+            jae.setTime(t.doubleValue());
 
         JsonNode v = o.get("v");
         if (v != null && v.isNumber())
