+++ leshan-core/src/main/java/org/eclipse/leshan/core/node/codec/senml/LwM2mNodeSenMLDecoder.java
@@ -171,12 +171,12 @@ public class LwM2mNodeSenMLDecoder implements TimestampedNodeDecoder, MultiNodeD
             SenMLPack pack = decoder.fromSenML(content);
 
             // Resolve records & Group it by time-stamp
-            Map<Float, Collection<LwM2mResolvedSenMLRecord>> recordsByTimestamp = groupRecordByTimestamp(
+            Map<Double, Collection<LwM2mResolvedSenMLRecord>> recordsByTimestamp = groupRecordByTimestamp(
                     pack.getRecords(), path);
 
             // Fill time-stamped nodes collection
             List<TimestampedLwM2mNode> timestampedNodes = new ArrayList<>();
-            for (Entry<Float, Collection<LwM2mResolvedSenMLRecord>> entryByTimestamp : recordsByTimestamp.entrySet()) {
+            for (Entry<Double, Collection<LwM2mResolvedSenMLRecord>> entryByTimestamp : recordsByTimestamp.entrySet()) {
                 LwM2mNode node = parseRecords(entryByTimestamp.getValue(), path, model, nodeClass);
                 // add time-stamped node
                 timestampedNodes.add(new TimestampedLwM2mNode(entryByTimestamp.getKey(), node));
@@ -356,11 +356,11 @@ public class LwM2mNodeSenMLDecoder implements TimestampedNodeDecoder, MultiNodeD
      * @return a sorted map (timestamp => collection of record) order by descending time-stamp (most recent one at first
      *         place). If null time-stamp (meaning no time information) exists it always at first place.
      */
-    private SortedMap<Float, Collection<LwM2mResolvedSenMLRecord>> groupRecordByTimestamp(List<SenMLRecord> records,
+    private SortedMap<Double, Collection<LwM2mResolvedSenMLRecord>> groupRecordByTimestamp(List<SenMLRecord> records,
             LwM2mPath requestPath) throws SenMLException {
-        SortedMap<Float, Collection<LwM2mResolvedSenMLRecord>> result = new TreeMap<>(new Comparator<Float>() {
+        SortedMap<Double, Collection<LwM2mResolvedSenMLRecord>> result = new TreeMap<>(new Comparator<Double>() {
             @Override
-            public int compare(Float o1, Float o2) {
+            public int compare(Double o1, Double o2) {
                 // null at first place
                 if (o1 == null && o2 == null)
                     return 0;
@@ -368,7 +368,7 @@ public class LwM2mNodeSenMLDecoder implements TimestampedNodeDecoder, MultiNodeD
                     return -1;
                 if (o2 == null)
                     return 1;
-                return Float.compare(o2, o1);
+                return Double.compare(o2, o1);
             }
         });
 
@@ -400,7 +400,7 @@ public class LwM2mNodeSenMLDecoder implements TimestampedNodeDecoder, MultiNodeD
         // Ensure there is at least one entry for null timestamp
         if (result.isEmpty()) {
             Collection<LwM2mResolvedSenMLRecord> emptylist = Collections.emptyList();
-            result.put((Float) null, emptylist);
+            result.put(null, emptylist);
         }
         return result;
     }
